+++ app/src/main/java/edu/agh/susgame/front/ui/components/game/map/components/drawers/EdgeDrawer.kt
@@ -1,4 +1,4 @@
-package edu.agh.susgame.front.ui.components.game.components.map.components.drawers
+package edu.agh.susgame.front.ui.components.game.map.components.drawers
 
 import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.layout.fillMaxSize
