+++ app/src/main/java/edu/agh/susgame/front/ui/components/game/map/components/elements/NodeInfoComp.kt
@@ -1,4 +1,4 @@
-package edu.agh.susgame.front.ui.components.game.components.map.components.elements
+package edu.agh.susgame.front.ui.components.game.map.components.elements
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
