+++ app/src/main/java/edu/agh/susgame/front/ui/components/menu/components/lobby/LobbyView.kt
@@ -44,4 +44,3 @@ fun LobbyView(
         }
     }
 }
-
