+++ app/src/main/java/edu/agh/susgame/front/ui/components/menu/navigation/MenuNavigationHost.kt
@@ -9,7 +9,7 @@ import edu.agh.susgame.front.navigation.MenuRoute
 import edu.agh.susgame.front.service.interfaces.GameService
 import edu.agh.susgame.front.service.interfaces.LobbyService
 import edu.agh.susgame.front.service.interfaces.ServerMapProvider
-import edu.agh.susgame.front.ui.components.game.components.map.GameView
+import edu.agh.susgame.front.ui.components.game.map.GameView
 import edu.agh.susgame.front.ui.components.menu.components.createlobby.CreateLobbyView
 import edu.agh.susgame.front.ui.components.menu.components.lobby.LobbyView
 import edu.agh.susgame.front.ui.components.menu.components.mainmenu.MainMenuView
